+++ kwimage/__init__.py
@@ -9,6 +9,31 @@ mkinit ~/code/kwimage/kwimage/structs/__init__.py --relative -w --nomod
 mkinit ~/code/kwimage/kwimage/__init__.py --relative --nomod  -w
 """
 
+import ubelt as ub
+try:
+    import cv2  # NOQA
+except Exception:
+    msg = ub.paragraph(
+        '''
+        The kwimage module failed to import the cv2 module.  This may be due to
+        an issue https://github.com/opencv/opencv-python/issues/467 which
+        prevents us from marking cv2 as package dependency.
+        To work around this we require that the user install this package with
+        one of the following extras tags:
+        `pip install kwimage[graphics]` xor
+        `pip install kwimage[headless]`.
+
+        Alternatively, the user can directly install the cv2 package as a post
+        processing step via:
+        `pip install opencv-python-headless` or
+        `pip install opencv-python`.
+
+        We appologize for this issue and hope this documentation is sufficient.
+        ''')
+    import warnings
+    warnings.warn(msg)
+
+
 __version__ = '0.7.2'
 
 from .algo import (available_nms_impls, daq_spatial_nms, non_max_supression,)